Make mapped virtual path's case insensitive
Noticed when using "MapPath" that the path seems to be case sensitive, when it shouldn't be:

                app.UseImageflow(new ImageflowMiddlewareOptions()
                    .SetMapWebRoot(false)
                    .MapPath("/vehicle", Path.Combine(dir, @"Inventory\Vehicle"))
                );

Calls like "https://localhost:44315/Vehicle/image.jpg?width=101" won't work because of the capital "V" in the virtual path. Changing it to a lowercase "v" will then produce the image. Should work either way or at least be configurable.

Note that this code change is untested, so you may need to verify it.